pyproject.toml
# Build System
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Project Metadata
[project]
name = "mcpguard"
version = "0.1.0"
description = "AI-powered security scanner for MCP servers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "pydantic-ai[mcp]>=1.70",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "logfire>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.2.0"
]

# Entry Point
[project.scripts]
mcpguard = "mcpguard.cli:app"

# Dev dependencies
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0"
]

# Hatchling Build Config (Critical for src layout)
[tool.hatch.build.targets.wheel]
packages = ["src/mcpguard"]

# Ruff Config
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

# Mypy Config
[tool.mypy]
python_version = "3.10"
strict = true

# Pytest Config
[tool.pytest.ini_options]
addopts = "--cov=mcpguard --cov-report=term-missing"
asyncio_mode = "auto"
testpaths = ["tests"]
